+++ src/pages/auth/hooks/index.ts
@@ -1,3 +1,2 @@
-export { useRestorePasswordForm } from "./use-restore-password-form.ts";
 export { useSignInForm } from "./use-sign-in-form.ts";
 export { useSignUpForm } from "./use-sign-up-form.ts";
